--- RealmLoginKit Android/library/src/main/java/io/realm/realmloginkit/Constants.java
@@ -1,9 +1,0 @@
-package io.realm.realmloginkit;
-
-public class Constants {
-    public static final String KEY_DARK_MODE = "DARK_MODE";
-    public static final String KEY_APP_TITLE = "APP_TITLE";
-
-    public static final int REQUEST_CODE_LOGIN_KIT = 200;
-    public static final int RESULT_CODE_OK = 1;
-}
