+++ RealmLoginKit Android/library/src/main/java/io/realm/realmloginkit/LoginKit.java
@@ -3,6 +3,9 @@ package io.realm.realmloginkit;
 import android.app.Activity;
 import android.content.Intent;
 
+import io.realm.realmloginkit.activity.RealmLoginActivity;
+import io.realm.realmloginkit.util.Constants;
+
 public class LoginKit {
     private Activity context;
     private Intent intent;
@@ -10,15 +13,8 @@ public class LoginKit {
     private LoginKit() {
     }
 
-    public static LoginKit loginKit(Activity context) {
-        LoginKit loginKit = new LoginKit();
-        loginKit.context = context;
-        loginKit.intent = new Intent(context, RealmLoginActivity.class);
-        return loginKit;
-    }
-
-    public LoginKit setDarkMode(boolean isDarkMode) {
-        intent.putExtra(Constants.KEY_DARK_MODE, isDarkMode);
+    public LoginKit setDarkMode(boolean darkMode) {
+        intent.putExtra(Constants.KEY_DARK_MODE, darkMode);
         return this;
     }
 
@@ -27,7 +23,20 @@ public class LoginKit {
         return this;
     }
 
+    public LoginKit setServerUri(String serverUri, boolean hideUri) {
+        intent.putExtra(Constants.KEY_SERVER_URI, serverUri);
+        intent.putExtra(Constants.KEY_HIDE_SERVER_URI, hideUri);
+        return this;
+    }
+
     public void logIn() {
-        context.startActivityForResult(intent, Constants.REQUEST_CODE_LOGIN_KIT);
+        context.startActivityForResult(intent, Constants.REQUEST_CODE_LOGIN);
+    }
+
+    public static LoginKit loginKit(Activity context) {
+        LoginKit loginKit = new LoginKit();
+        loginKit.context = context;
+        loginKit.intent = new Intent(context, RealmLoginActivity.class);
+        return loginKit;
     }
 }
