+++ RealmLoginKit Android/library/src/main/java/io/realm/realmloginkit/util/Constants.java
@@ -1,0 +1,11 @@
+package io.realm.realmloginkit.util;
+
+public class Constants {
+    public static final String KEY_DARK_MODE = "DARK_MODE";
+    public static final String KEY_APP_TITLE = "APP_TITLE";
+    public static final String KEY_SERVER_URI = "SERVER_URI";
+    public static final String KEY_HIDE_SERVER_URI = "HIDE_SERVER_URI";
+
+    public static final int REQUEST_CODE_LOGIN = 200;
+    public static final int RESULT_CODE_LOGIN_OK = 1;
+}
