+++ RealmLoginKit Android/library/src/main/res/layout/activity_register.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="io.realm.realmloginkit.RealmLoginActivity">
+    tools:context="io.realm.realmloginkit.activity.RealmLoginActivity">
 
     <LinearLayout
         android:id="@+id/login_panel"
